Lesson 2.13: moved ForecastFragment to its own file
