+++ app/src/main/java/it/iochiamo/android/mysunshine/MainActivity.java
@@ -1,22 +1,7 @@
 package it.iochiamo.android.mysunshine;
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.support.v7.app.AppCompatActivity;
-import android.util.Log;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.ArrayAdapter;
-import android.widget.ListView;
-
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
-import java.net.URL;
-import java.util.ArrayList;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -35,98 +20,5 @@ public class MainActivity extends AppCompatActivity {
         }
     }
 
-    public static class ForecastFragment extends Fragment {
-
-
-        public ForecastFragment() {
-        }
-
-        @Override
-        public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                                 Bundle savedInstanceState) {
-            View rootView = inflater.inflate(R.layout.fragment_main, container, false);
-            // Inflate the layout for this fragment
-            ArrayList<String> weekForecastArray = new ArrayList<>();
-
-            int i = 0;
-            while (i < 40) {
-                weekForecastArray.add("Weather Forecast for day " + Integer.toString(i + 1));
-                i++;
-            }
-            ArrayAdapter<String> mForecastAdapter;
-
-            mForecastAdapter =
-                    new ArrayAdapter<String>(
-                            getContext(), // The current context (this activity)
-                            R.layout.list_item_forecast, // The name of the layout ID.
-                            R.id.list_item_forecast_textview, // The ID of the textview to populate.
-                            weekForecastArray);
-
-            ListView listView = (ListView) rootView.findViewById(R.id.listview_forecast);
-
-            listView.setAdapter(mForecastAdapter);
-
-// These two need to be declared outside the try/catch
-// so that they can be closed in the finally block.
-            HttpURLConnection urlConnection = null;
-            BufferedReader reader = null;
-
-// Will contain the raw JSON response as a string.
-            String forecastJsonStr = null;
-
-            try {
-                // Construct the URL for the OpenWeatherMap query
-                // Possible parameters are available at OWM's forecast API page, at
-                // http://openweathermap.org/API#forecast
-                URL url = new URL("http://api.openweathermap.org/data/2.5/forecast?units=metric&id=3169070&appid=99f0e966b8f6a08e87b59b9336428b19&mode=json&cnt=3");
-
-                // Create the request to OpenWeatherMap, and open the connection
-                urlConnection = (HttpURLConnection) url.openConnection();
-                urlConnection.setRequestMethod("GET");
-                urlConnection.connect();
-
-                // Read the input stream into a String
-                InputStream inputStream = urlConnection.getInputStream();
-                StringBuffer buffer = new StringBuffer();
-                if (inputStream == null) {
-                    // Nothing to do.
-                    forecastJsonStr = null;
-                }
-                reader = new BufferedReader(new InputStreamReader(inputStream));
-
-                String line;
-                while ((line = reader.readLine()) != null) {
-                    // Since it's JSON, adding a newline isn't necessary (it won't affect parsing)
-                    // But it does make debugging a *lot* easier if you print out the completed
-                    // buffer for debugging.
-                    buffer.append(line + "\n");
-                }
-
-                if (buffer.length() == 0) {
-                    // Stream was empty.  No point in parsing.
-                    forecastJsonStr = null;
-                }
-                forecastJsonStr = buffer.toString();
-            } catch (IOException e) {
-                Log.e("ForecastFragment", "Error ", e);
-                // If the code didn't successfully get the weather data, there's no point in attempting
-                // to parse it.
-                forecastJsonStr = null;
-            } finally{
-                if (urlConnection != null) {
-                    urlConnection.disconnect();
-                }
-                if (reader != null) {
-                    try {
-                        reader.close();
-                    } catch (final IOException e) {
-                        Log.e("ForecastFragment", "Error closing stream", e);
-                    }
-                }
-            }
-
-            return rootView;
-        }
-    }
 
 }
